Cleanup

+++ example/OLED_SSD1306.c
@@ -71,7 +71,7 @@ int8_t OLED_1306_InvertColors(){
 	}
 }
 
-int8_t OLED_1306_DrawString(int x, int y, const char* str, uint8_t font_size, uint8_t color){
+int8_t OLED_1306_DrawString(uint8_t x, uint8_t y, const char* str, uint8_t font_size, uint8_t color){
 	int8_t status = 0;
 
 	int x_tmp = x;
@@ -88,7 +88,7 @@ int8_t OLED_1306_DrawString(int x, int y, const char* str, uint8_t font_size, ui
 	return status;
 }
 
-int8_t OLED_1306_DrawCharacter(uint16_t x, uint16_t y, char chr, uint8_t font_size, uint8_t color){
+int8_t OLED_1306_DrawCharacter(uint8_t x, uint8_t y, char chr, uint8_t font_size, uint8_t color){
 	if(chr > 0x7E) return 255; // chr > '~'
 
 	int8_t status = 0;
